requirements.txt
python-dotenv >= 0.0.5
pynput == 1.7.7